file(#255): 시간 표기 관련 util 함수 별도 파일로 분리
